% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bracken_clean.R
\name{bracken_clean}
\alias{bracken_clean}
\title{Clean Bracken}
\usage{
bracken_clean(data, min_abund = 0.01)
}
\arguments{
\item{data}{input data (can be piped from readR etc.)}

\item{min_abund}{option adjust minimum abundance (default 0.01)}
}
\description{
This function accepts a bracken summary file from the bracken python script
and returns a cleaned data frame for further analyses.
}
\examples{
bracken_clean(data = "bracken_s")
data \%>\% bracken_clean(min_abund=0.05)
}
\keyword{MEP}
