% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_prefix.R
\name{get_prefix}
\alias{get_prefix}
\title{Get prefixes}
\usage{
get_prefix(names, sep = "-", HMP = FALSE)
}
\arguments{
\item{names}{list of names to pull prefixes}

\item{sep}{character string to split names (e.g. "-")}

\item{HMP}{option to add HMP prefix SRS to list if HMP=TRUE, defaults to HMP=FALSE.}
}
\description{
This function accepts a list of characters and returns
a list of unique prefixes.
}
\examples{
get_prefix(colnames(RPKM_m), sep="-")
get_prefix(colnames(RPKM_m), sep="-", HMP= TRUE)
}
\keyword{MEP}
