% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pt_heat.R
\name{pt_heat}
\alias{pt_heat}
\title{Patient pheatmaps}
\usage{
pt_heat(
  matrix_input,
  prefixes,
  rowmeta = NULL,
  colmeta = NULL,
  meta_colors = NULL,
  param_string = NULL
)
}
\arguments{
\item{prefixes}{list of prefixes (see get_prefix)}

\item{matrix}{numerix matrix}

\item{colannotation}{dataframe with column annotation metadata}

\item{rowannotation}{dataframe with row annotation metadata}

\item{colors}{list of hex color values (see pheatmap(annotation_colors))}

\item{options}{string of comma separated pheatmap options to pass (e.g. cluster_rows, scale...)}
}
\description{
This function accepts a numeric matrix (typically features as rows,
and samples as columns) list of patient prefixes and plots heatmaps
for data by patient.
}
\details{
To add column & row metadata with color specifications, these can
be supplied as optional parameters.  Metadata annotations must be
stored as data frames with rownames that correspond with matrix rownames.
}
\examples{
get_prefix(colnames(RPKM_m), sep="-", HMP= TRUE)
}
\keyword{MEP}
